requirements.txt
flask==2.0.1
paho-mqtt==1.6.1
gunicorn==20.1.0
werkzeug==2.0.3
flask-socketio==5.3.6
requests==2.31.0
python-engineio==4.9.0
python-socketio==5.11.2
eventlet==0.35.1 